bronze/requirements.txt
# Minimal deps for GCS upload and BigQuery load
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
pandas>=2.0.0
